+++ src/index.ts
@@ -1,9 +1,10 @@
-// Bootstrap source mapping
-import * as sourceMapper from 'source-map-support';
-
-sourceMapper.install({
-	environment: 'node'
-});
+// Source mapping
+if (!process[Symbol.for('ts-node.register.instance')]) {
+	// eslint-disable-next-line @typescript-eslint/no-var-requires
+	require('source-map-support').install({
+		environment: 'node'
+	});
+}
 
 // Enter application
 import './main';